usm: Move Supported() helpers out of http package
The minimum requirements for USM are common to several protocols
(although a couple have separate, higher requirements), so move the
helper functions to the usmconfig package instead of the http package.

Also, this allows us to remove the dependency on the pkg/ebpf package
from the agent binary.  The agent binary got this dependencty via the
http package with the recent change to build kernel_requirements.go from
that package on linux instead of linux_bpf.
