+++ pkg/network/usm/ebpf_ssl_test.go
@@ -25,7 +25,7 @@ func testArch(t *testing.T, arch string) {
 	cfg := config.New()
 	cfg.EnableNativeTLSMonitoring = true
 
-	if !http.TLSSupported(cfg) {
+	if !usmconfig.TLSSupported(cfg) {
 		t.Skip("shared library tracing not supported for this platform")
 	}
